+++ opentelemetry-sdk/src/metrics/periodic_reader_with_async_runtime.rs
@@ -449,7 +449,7 @@ mod tests {
         runtime, Resource,
     };
     use opentelemetry::metrics::MeterProvider;
-    use std::sync::mpsc;
+    use std::sync::{mpsc, Arc, Mutex};
 
     #[test]
     fn collection_triggered_by_interval_tokio_current() {
@@ -492,8 +492,12 @@ mod tests {
     #[test]
     fn unregistered_collect() {
         // Arrange
-        let exporter = InMemoryMetricExporter::default();
-        let reader = PeriodicReader::builder(exporter.clone(), runtime::Tokio).build();
+        let metrics = Arc::new(Mutex::new(Vec::new()));
+        let exporter = InMemoryMetricExporter::builder()
+            .with_metrics(metrics.clone())
+            .build();
+
+        let reader = PeriodicReader::builder(exporter, runtime::Tokio).build();
         let mut rm = ResourceMetrics {
             resource: Resource::empty(),
             scope_metrics: Vec::new(),
@@ -513,8 +517,13 @@ mod tests {
         RT: crate::runtime::Runtime,
     {
         let interval = std::time::Duration::from_millis(1);
-        let exporter = InMemoryMetricExporter::default();
-        let reader = PeriodicReader::builder(exporter.clone(), runtime)
+
+        let metrics = Arc::new(Mutex::new(Vec::new()));
+        let exporter = InMemoryMetricExporter::builder()
+            .with_metrics(metrics.clone())
+            .build();
+
+        let reader = PeriodicReader::builder(exporter, runtime)
             .with_interval(interval)
             .build();
         let (sender, receiver) = mpsc::channel();